+++ models/schemas/role.js
@@ -11,15 +11,23 @@ let schema = new Schema({
         required: true,
         unique: true
     },
-
     desc: {
         type: String,
         default: ''
     },
-
+    // 未來可能是寄信可以知道層次，目前系統管理者是 999
+    level: {
+        type: Number,
+        default: 0
+    },
     Policies: [{
         type: Schema.Types.ObjectId,
         ref: 'Policy'
+    }],
+    // 可以審此角色的角色
+    SuperiorRoles: [{
+        type: Schema.Types.ObjectId,
+        ref: 'Role'
     }],
 
     isTrashed: {